Building DestinationClientHandler:
Created structs for CTMP Messages which will be used later by the Messages.

Turned DestinationClientHandler into a singleton for temporary construction(until a better design pattern is found).

Added destination clients to the Destination Client Handler.

Added a queue of shared pointers to the Destination client.
Added locks(mutexs) to the Destination Client which caused an issue with the hashmap, so that has been converted to shared pointers aswell.